refactor public API to expose collect methods
